Fix potential dangling pointer in marshalObject
marshalObject returns a new Go object that holds a copy of the C GObject
pointer retrieved from the GValue, but the reference count on the C
GObject is not increased. The GValue created in GoValue() is no longer
in scope after GoValue returns, and can be freed by the garbage
collector. At this point, if no non-Go code holds a reference to the C
GObject, then the C GObject reference count will drop to zero, the C
GObject will be disposed, and the Go object pointer to it will become a
dangling pointer.

Fix this by calling the `glib.Take` function in marshalObject. This
increments the reference count for the C GObject, and adds a finaliser
to the Go Object which will decrement the C reference count when the Go
object is garbage collected.